+++ app/Http/Controllers/Api/TenantRoleController.php
@@ -8,6 +8,7 @@ use App\Models\TenantRole;
 use App\Services\RoleService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TenantRoleController extends Controller
 {
@@ -21,6 +22,8 @@ class TenantRoleController extends Controller
     public function index(Request $request): JsonResponse
     {
         $tenantId = $request->user()->tenant_id;
+        Gate::authorize('viewAny', TenantRole::class);
+
         $roles = TenantRole::where('tenant_id', $tenantId)->get();
         return response()->json($roles);
     }
@@ -31,6 +34,7 @@ class TenantRoleController extends Controller
     public function store(StoreTenantRoleRequest $request): JsonResponse
     {
         $cridentials = $request->validated();
+        Gate::authorize('create', TenantRole::class);
 
         $tenant = $request->user()->tenant;
 
@@ -47,6 +51,8 @@ class TenantRoleController extends Controller
      */
     public function destroy(TenantRole $tenantRole): JsonResponse
     {
+        Gate::authorize('delete', $tenantRole);
+        
         $this->roleService->deleteTenantRole($tenantRole);
         return response()->json(null, 204);
     }
